fix: make silent failures visible; add post-run display review
The recommendation crew had been failing on every run since at least
2026-08-02 — OpenAI rejects the regex lookaround Pydantic emits for Decimal
fields, so output_pydantic=InvestmentRecommendation 400'd. It was caught,
logged at WARNING, and the run still reported "completed". With
LOG_LEVEL=ERROR the warning was never written, so the only remaining
detection mechanism was a human noticing two history tiles looked different.

Root fix: target_price/stop_loss are float, not Decimal. A test asserts the
generated JSON schema stays lookaround-free.

Stop the class of bug, not just this instance:

- run_review.py checks every completed run against the *display contract* —
  the exact keys list_reports() and dashboard.js read — then logs, stores the
  verdict in the status marker, and shows "⚠ Incomplete data" on the tile.
- The flow now records degradations: caught-and-worked-around failures
  (recommendation, report narrative, HTML render, chart data) travel back in
  the result instead of living only in the log. They deliberately do NOT fail
  the run — a partial report still beats none — they just stop being
  invisible. Folded into the review, because a failed stage can leave a stale
  artifact from a previous run that every file check would call healthy.
- The CLI prints them too; it has no tile to show them on.
- Un-silence three handlers that hid real damage: a failure in post-run
  alerts meant alerts were never dispatched and looked exactly like "none
  were due"; the others lost the diff baseline and rec_history points.
- LOG_LEVEL: ERROR -> INFO in .env, so the warnings exist at all.

Also fix a genuine race (pre-existing; it made test_cancel_marks_job_aborted
fail ~50% of runs): _run() set job.state terminal inside the try, while the
status marker was written later in finally. The UI polls every second, so it
could observe "completed" and open a report whose marker did not exist yet.
The outcome is now accumulated locally and published only after the marker is
on disk. 6/6 clean runs, previously 2-of-4.

Tests: 513 passed (31 new).

+++ comprehensive_stock_analysis/src/stock_analysis/models/stock_data.py
@@ -173,8 +173,16 @@ class InvestmentRecommendation(BaseModel):
     recommendation: RecommendationType = Field(
         ..., description="Investment recommendation"
     )
-    target_price: Optional[Decimal] = Field(None, description="Target price")
-    stop_loss: Optional[Decimal] = Field(None, description="Stop loss price")
+    # float, not Decimal, and this is load-bearing: this model is handed to
+    # CrewAI as `output_pydantic`, which becomes an OpenAI structured-output
+    # schema. Pydantic renders Decimal as anyOf[number, string] where the
+    # string branch carries the pattern "^(?!^[-+.]*$)...", and OpenAI rejects
+    # the (?!...) lookaround with "Invalid JSON schema: regex lookaround is not
+    # supported". That 400 failed the recommendation crew on every run, so no
+    # <SYM>_investment_recommendation.json was written and every report fell
+    # back to an N/A rating. Keep these JSON-schema-safe.
+    target_price: Optional[float] = Field(None, description="Target price")
+    stop_loss: Optional[float] = Field(None, description="Stop loss price")
     time_horizon: str = Field(..., description="Investment time horizon")
     risk_level: RiskLevel = Field(..., description="Risk level")
     confidence: float = Field(..., description="Confidence level (0-1)")
@@ -208,7 +216,7 @@ class InvestmentRecommendation(BaseModel):
             match = re.search(r"[-+]?\d[\d,]*(?:\.\d+)?", v)
             if match is None:
                 return None
-            return Decimal(match.group(0).replace(",", ""))
+            return float(match.group(0).replace(",", ""))
         return v
 
     @field_validator("confidence")
